refactor(ParticipantDto): cards와 score필드 두가지를 추가로 갖도록 변경
- HandDto 제거
